restructured the gameobjects, sprites, shields and invaders.
VisualGameObject was split into a core VisualGameObject and a Sprite object that handles rendering. both will support the decorator pattern to modify their properties, with AnimatedSprite being an example of this.

the shields and invaders have both been consolidated, so that there is no longer multiple files to contain the additional rules. ShieldList.js, InvaderController.js, InvaderWaves.js have all been deleted and their contents relocated to either Shield.js or Invader.js, as needed.

while doing this, i also patched a bug with the generators where the health bars would not show properly when health was maxed out after losing health and restoring it.